+++ CHANGELOG.md
@@ -9,12 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Fixed
+
+## [0.2.5] - 2023-08-17
+
+### Added
+
 - Fast alternative to partial trace (`Statevec.remove_qubit`) for a separable (post-measurement) qubit (#73)
 
 ### Changed
 
 - `StatevectorBackend` now uses `Statevec.remove_qubit` after each measurement, instead of performing `ptrace` after multiple measurements, for better performance. This keeps the result exactly the same (#73)
-
+- bump dependency versions for docs build (#77)
 
 ## [0.2.4] - 2023-07-06
 ### Added